Reorganize config manager packages in agent
