+++ cmd/agent/app/reporter/grpc/collector_proxy.go
@@ -25,6 +25,7 @@ import (
 	"google.golang.org/grpc/resolver/manual"
 
 	"github.com/jaegertracing/jaeger/cmd/agent/app/httpserver"
+	mGrpc "github.com/jaegertracing/jaeger/cmd/agent/app/httpserver/grpc"
 	aReporter "github.com/jaegertracing/jaeger/cmd/agent/app/reporter"
 )
 
@@ -57,7 +58,7 @@ func NewCollectorProxy(o *Options, mFactory metrics.Factory, logger *zap.Logger)
 	return &ProxyBuilder{
 		conn:     conn,
 		reporter: aReporter.WrapWithMetrics(NewReporter(conn, logger), grpcMetrics),
-		manager:  httpserver.WrapWithMetrics(NewSamplingManager(conn), grpcMetrics)}, nil
+		manager:  httpserver.WrapWithMetrics(mGrpc.NewSamplingManager(conn), grpcMetrics)}, nil
 }
 
 // GetReporter returns Reporter
